requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
ibm_db==3.2.6
idna==3.10
mysql-connector-python==9.2.0
oracledb==3.0.0
psycopg==3.2.5
psycopg-binary==3.2.5
psycopg-pool==3.2.6
pycparser==2.22
pymssql==2.3.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
